Rename Data cleaning for classicification (3-Table1-1).png to Data cleaning for classification (3-Table1-1).png
The following table offers a thorough summary of the four datasets that were utilized for classification tasks. The quantity of patterns, their characteristics, and how they are distributed among the two classifications define each dataset. For the purpose of creating and assessing classification models, it is necessary to comprehend the size and complexity of each dataset.

I made a pie chart and a bar graph to help visualize the data and make it easier to understand.